refactor(path-matcher): split registration.ts into a registration/ folder (no behavior change)
registration.ts had grown to 1296 lines mixing error factories, trie-node
construction, recursive trie insertion, build-part compilation, and
orchestration. Split into 6 cohesive, cycle-free modules under registration/
(dependency order context <- errors <- trieNodes <- trie <- buildParts <- index),
each <=354 lines.

Pure code move: function bodies are byte-identical (node-spliced, not retyped —
preserves the regex/template/unicode message literals); only per-file imports +
`export` on cross-module functions were added. The one external entry
(registerNode, imported by SegmentMatcher) is re-exported from
registration/index.ts, so `from "./registration"` resolves unchanged.

No behavior change: path-matcher 497 unit / 100% cov / 97 property / 20 stress,
route-tree 240, core 2487 — all green; knip clean on registration/ exports. The
published bundle is byte-identical, so no changeset (path-matcher is private,
bundled into core; file layout does not affect the bundle).
